+++ Frontend/fe_vue/src/view/roadmap.vue
@@ -13,9 +13,7 @@
         </div>
 
     </div>
-    <button @click="getSubData" v-if="!roadMapShowClicked">
-        temp btn for getting data
-    </button>
+
 
 </template>
 
@@ -26,11 +24,10 @@ export default {
 
     data() {
         return {
-            roadMapShowClicked: false,
             isCardOn: false,
             success: false,
             subDataSuccess:false,
-            roadmapId: "temp",
+            roadmapId: "1",
             roadmapDetail: [],
 
             curSubjectId : "",
@@ -45,35 +42,6 @@ export default {
         setSubIdFromData(subid) {
             this.curSubjectId = subid;
         },
-        /** 
-         * roadmapDetail : 
-         * "subjects" : 
-    {
-        //[name, id]
-        "1" : [["기초프로그래밍", 1001]] ,
-        "2" : [["파이썬 기초", 2001], ["자바의 기초",2002]],
-        "3" : [["자료구조와 알고리즘", 3001], ["운영체제",3002]],
-        "4" : [["김영한의 스프링", 4001 ]]
-    }
-         * 
-         * 
-         * subdetail : 
-         * {
-   "success" : "true",
-   "courses" : 
-   [ 
-                {
-            "subjectId":123, 
-            "courseId":123, 
-            "courseName":"name", 
-            "thumbnailUrl":"url", 
-            "courseUrl":"url", 
-            "explain": "hello"
-        } , { ... } 
-    ]
-   
-}
-         */
         getCourseData(subject) {
             this.isCardOn = true;
             console.log("subject : " +subject);
@@ -91,20 +59,7 @@ export default {
                     console.log(error)
                 })
         },
-        getSubData: function () {
-            this.roadMapShowClicked = true;
-            var vm = this;
-            axios.get('https://backend.devroad.site/' + 'api/roadmap/' + this.roadmapId)
-                .then(response => {
-                    console.log(response);
-                    vm.isSuccess = response.data.isSuccess;
-                    vm.roadmapDetail = response.data.roadmapDetail;
-
-                })
-                .catch(function (error) {
-                    console.log(error)
-                })
-        }
+       
     }
